backend/requirements.txt
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
gunicorn==20.1.0
ibm-db==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.0
Werkzeug==2.2.2
